refactor: Restructure packages and add abstract MenuView & BoardCreatorView
